Terrain/RasterTileCache: limit file offsets to 32 bit
We were using "long", which was 32 bit on ARM, and 64 bit on AMD64.
We don't need to support >4GB maps, our map files are usually smaller
than 10 MB.

+++ src/Terrain/RasterTile.cpp
@@ -321,7 +321,7 @@ RasterTileCache::Reset()
 
 gcc_pure
 const RasterTileCache::MarkerSegmentInfo *
-RasterTileCache::FindMarkerSegment(long file_offset) const
+RasterTileCache::FindMarkerSegment(uint32_t file_offset) const
 {
   for (const MarkerSegmentInfo *p = segments.begin(); p < segments.end(); ++p)
     if (p->file_offset >= file_offset)
